test: moves swap, wallet and Trending specs to tests
