+++ tests/smoke/swap/unified-ui-wallet-actions.spec.ts
@@ -1,14 +1,14 @@
 'use strict';
 import { Mockttp } from 'mockttp';
-import { loginToApp } from '../../viewHelper';
-import TabBarComponent from '../../pages/wallet/TabBarComponent';
-import WalletActionsBottomSheet from '../../pages/wallet/WalletActionsBottomSheet';
-import FixtureBuilder from '../../../tests/framework/fixtures/FixtureBuilder';
-import { withFixtures } from '../../../tests/framework/fixtures/FixtureHelper';
-import { SmokeTrade } from '../../tags.js';
-import WalletView from '../../pages/wallet/WalletView';
-import Assertions from '../../../tests/framework/Assertions';
-import TestHelpers from '../../helpers.js';
+import { loginToApp } from '../../../e2e/viewHelper.ts';
+import TabBarComponent from '../../../e2e/pages/wallet/TabBarComponent.ts';
+import WalletActionsBottomSheet from '../../../e2e/pages/wallet/WalletActionsBottomSheet.ts';
+import FixtureBuilder from '../../framework/fixtures/FixtureBuilder.ts';
+import { withFixtures } from '../../framework/fixtures/FixtureHelper.ts';
+import { SmokeTrade } from '../../../e2e/tags.js';
+import WalletView from '../../../e2e/pages/wallet/WalletView.ts';
+import Assertions from '../../framework/Assertions.ts';
+import TestHelpers from '../../../e2e/helpers.js';
 
 describe(SmokeTrade('Trade: Unified UI Wallet Actions'), () => {
   let mockServer: Mockttp | undefined;
