+++ tests/smoke/trending/trending-browser.spec.ts
@@ -1,16 +1,16 @@
-import { SmokeWalletPlatform } from '../../tags';
-import { loginToApp, navigateToBrowserView } from '../../viewHelper';
-import FixtureBuilder from '../../../tests/framework/fixtures/FixtureBuilder';
-import { withFixtures } from '../../../tests/framework/fixtures/FixtureHelper';
-import { DappVariants } from '../../../tests/framework/Constants';
+import { SmokeWalletPlatform } from '../../../e2e/tags';
+import { loginToApp, navigateToBrowserView } from '../../../e2e/viewHelper.ts';
+import FixtureBuilder from '../../framework/fixtures/FixtureBuilder.ts';
+import { withFixtures } from '../../framework/fixtures/FixtureHelper.ts';
+import { DappVariants } from '../../framework/Constants.ts';
 import { Mockttp } from 'mockttp';
-import { Assertions } from '../../../tests/framework';
-import { setupRemoteFeatureFlagsMock } from '../../../tests/api-mocking/helpers/remoteFeatureFlagsHelper';
-import { TRENDING_API_MOCKS } from '../../../tests/api-mocking/mock-responses/trending-api-mocks';
-import { setupMockEvents } from '../../../tests/api-mocking/helpers/mockHelpers';
-import { remoteFeatureFlagTrendingTokensEnabled } from '../../../tests/api-mocking/mock-responses/feature-flags-mocks';
-import Browser from '../../pages/Browser/BrowserView';
-import TestDApp from '../../pages/Browser/TestDApp';
+import { Assertions } from '../../framework';
+import { setupRemoteFeatureFlagsMock } from '../../api-mocking/helpers/remoteFeatureFlagsHelper.ts';
+import { TRENDING_API_MOCKS } from '../../api-mocking/mock-responses/trending-api-mocks.ts';
+import { setupMockEvents } from '../../api-mocking/helpers/mockHelpers.ts';
+import { remoteFeatureFlagTrendingTokensEnabled } from '../../api-mocking/mock-responses/feature-flags-mocks.ts';
+import Browser from '../../../e2e/pages/Browser/BrowserView.ts';
+import TestDApp from '../../../e2e/pages/Browser/TestDApp.ts';
 
 describe(SmokeWalletPlatform('Trending Feature Browser Test'), () => {
   const testSpecificMock = async (mockServer: Mockttp) => {
