+++ tests/smoke/trending/trending-feed.spec.ts
@@ -1,17 +1,17 @@
-import { SmokeWalletPlatform } from '../../tags';
-import { loginToApp } from '../../viewHelper';
-import FixtureBuilder from '../../../tests/framework/fixtures/FixtureBuilder';
-import { withFixtures } from '../../../tests/framework/fixtures/FixtureHelper';
-import { setupRemoteFeatureFlagsMock } from '../../../tests/api-mocking/helpers/remoteFeatureFlagsHelper';
+import { SmokeWalletPlatform } from '../../../e2e/tags';
+import { loginToApp } from '../../../e2e/viewHelper.ts';
+import FixtureBuilder from '../../framework/fixtures/FixtureBuilder.ts';
+import { withFixtures } from '../../framework/fixtures/FixtureHelper.ts';
+import { setupRemoteFeatureFlagsMock } from '../../api-mocking/helpers/remoteFeatureFlagsHelper.ts';
 import { Mockttp } from 'mockttp';
-import TrendingView from '../../../tests/page-objects/Trending/TrendingView';
-import { TrendingViewSelectorsText } from '../../../tests/locators/Trending/TrendingView.selectors';
-import { TRENDING_API_MOCKS } from '../../../tests/api-mocking/mock-responses/trending-api-mocks';
-import { setupMockEvents } from '../../../tests/api-mocking/helpers/mockHelpers';
+import TrendingView from '../../page-objects/Trending/TrendingView.ts';
+import { TrendingViewSelectorsText } from '../../locators/Trending/TrendingView.selectors.ts';
+import { TRENDING_API_MOCKS } from '../../api-mocking/mock-responses/trending-api-mocks.ts';
+import { setupMockEvents } from '../../api-mocking/helpers/mockHelpers.ts';
 import {
   remoteFeatureFlagTrendingTokensEnabled,
   remoteFeatureFlagPredictEnabled,
-} from '../../../tests/api-mocking/mock-responses/feature-flags-mocks';
+} from '../../api-mocking/mock-responses/feature-flags-mocks.ts';
 
 describe(SmokeWalletPlatform('Trending Feed View All Navigation'), () => {
   const testSpecificMock = async (mockServer: Mockttp) => {
