+++ tests/smoke/wallet/helpers/musd-fixture.ts
@@ -1,14 +1,11 @@
 import FixtureBuilder, {
   type MusdFixtureOptions,
-} from '../../../../tests/framework/fixtures/FixtureBuilder';
+} from '../../../framework/fixtures/FixtureBuilder.ts';
 import { CHAIN_IDS } from '@metamask/transaction-controller';
 import { toChecksumHexAddress } from '@metamask/controller-utils';
-import { AnvilPort } from '../../../../tests/framework/fixtures/FixtureUtils';
-import { AnvilManager } from '../../../../tests/seeder/anvil-manager';
-import {
-  USDC_MAINNET,
-  MUSD_MAINNET,
-} from '../../../../tests/constants/musd-mainnet';
+import { AnvilPort } from '../../../framework/fixtures/FixtureUtils.ts';
+import { AnvilManager } from '../../../seeder/anvil-manager.ts';
+import { USDC_MAINNET, MUSD_MAINNET } from '../../../constants/musd-mainnet.ts';
 
 const USDC_DECIMALS = 6;
 const MUSD_DECIMALS = 6;
